+++ mapsforge-core/src/main/java/org/mapsforge/core/util/Parameters.java
@@ -17,6 +17,11 @@ package org.mapsforge.core.util;
 public final class Parameters {
 
     /**
+     * If true the <code>MapViewPosition2</code> will be used instead of default <code>MapViewPosition</code>.
+     */
+    public static boolean MAP_VIEW_POSITION2 = false;
+
+    /**
      * Maximum buffer size for map files.
      */
     public static int MAXIMUM_BUFFER_SIZE = 8000000;